CODE-2746 CR of 1/3 or less always return 1/2 Fix design bug. CR is now internally being represented as Integer instead of Float, with fractional CRs being represented by negative values.
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@25450 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/core/PCTemplate.java
@@ -61,12 +61,12 @@ public final class PCTemplate extends PObject implements TemplateFacade, ChooseD
 	 * 
 	 * @return a Challenge Rating adjustment
 	 */
-	public float getCR(final int level, final int hitdice)
+	public Integer getCR(final int level, final int hitdice)
 	{
-		float localCR = getSafe(ObjectKey.CR_MODIFIER).floatValue();
+		Integer localCR = getSafe(ObjectKey.CR_MODIFIER).intValue();
 		for (PCTemplate pct : getConditionalTemplates(level, hitdice))
 		{
-			localCR += pct.getSafe(ObjectKey.CR_MODIFIER).floatValue();
+			localCR += pct.getSafe(ObjectKey.CR_MODIFIER).intValue();
 		}
 		return localCR;
 	}
